pyproject.toml
[project]
name = "connectonion-gmail-agent"
version = "0.2.0"
description = "Multi-agent Gmail management using ConnectOnion"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Sreekar Reddy", email = "esreekarreddy@gmail.com"}
]
keywords = ["connectonion", "gmail", "ai-agent", "multi-agent", "email"]

dependencies = [
    "connectonion>=0.4.1",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "tox>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gmail-agent = "cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "cli", "schemas", "plugins"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
